atualizações do dia 10 e noite de 11/06/2023

+++ Back/src/main/java/com/barbershowa/resource/ClienteResource.java
@@ -3,15 +3,19 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import com.barbershowa.model.Cliente;
+import com.barbershowa.repository.AgendamentoRepository;
 import com.barbershowa.repository.ClienteRepository;
 import java.net.URI;
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
+import java.util.Date;
 
 @CrossOrigin(origins = "http://localhost:4200")
 @RestController
@@ -20,6 +24,9 @@ public class ClienteResource {
 	
 	@Autowired
 	private ClienteRepository clienteRepository;
+	
+	@Autowired
+	private AgendamentoRepository agendamentoRepository;
 	
 	@GetMapping
 	public List<Cliente> list() {
@@ -51,10 +58,23 @@ public class ClienteResource {
 		return ResponseEntity.ok(cliente);
 	  }
 	  
+	@Transactional
 	@DeleteMapping("/{id}")
 	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void delete(@PathVariable Integer	id) {
-		clienteRepository.deleteById(id);
+	public void delete(@PathVariable Integer id) {
+		
+		Optional<Cliente> clienteOptional = clienteRepository.findById(id);
+	    if (clienteOptional.isPresent()) {
+	        Cliente cliente = clienteOptional.get();
+	        Date dataAtual = java.sql.Date.valueOf(LocalDate.now());
+	        // antes de apagar o cliente, apaga agendamentos dele atuais - não apaga antigos - apaga data igual e superior a data atual
+	        agendamentoRepository.deleteByClienteAndDatasGreaterThanEqual(cliente, dataAtual);
+	        // altera coluna cliente para id=31, na tabela agendamentos
+			agendamentoRepository.updateClienteIdByClienteId(31, id);
+			// Exclui o cliente - na tabela "cliente"
+			clienteRepository.deleteById(id);    
+	    }
+		
 	}
 	  
 }
